// src/app/interfaces/inspection-master.ts
export interface InspectionMaster {
    subscriberId?: number;
    inspStageId?: number;
    inspStageName?: string;
    subscriberName?: string;
    inspectionMasterId?: number;
    componentProductDrawNumber?: string;
    componentProductName?: string;
    componentProductNumber?: string;
    componentProductMaterial?: string;
    componentProductNotes?: string;
    inspectionType?: number;
    inspectionStage?: number;
}